+++ losses/base.py
@@ -172,84 +172,6 @@ class ModelSaver(TestListener):
         }, self.path)
 
 
-class Evaluator(TrainingListener):
-    
-    def __init__(self, device, loader, metric, fit_metric=False, batch_transforms=[], callbacks=[]):
-        super(Evaluator, self).__init__()
-        self.device = device
-        self.loader = loader
-        self.metric = metric
-        self.fit_metric = fit_metric
-        self.batch_transforms = batch_transforms
-        self.callbacks = callbacks
-        self.feat_train, self.y_train = None, None
-        self.best_metric, self.best_epoch = 0, -1
-    
-    def _eval(self, model):
-        model.eval()
-        feat_test, logits_test, y_test = [], [], []
-        for cb in self.callbacks:
-            cb.on_before_test()
-        with torch.no_grad():
-            for i in range(self.loader.nbatches()):
-                x, y = next(self.loader)
-
-                # Apply custom transformations to the batch before feeding the model
-                for transform in self.batch_transforms:
-                    x, y = transform(x, y)
-                
-                # Feed Forward
-                feat, logits = model(x, y)
-                feat = feat.detach().cpu().numpy()
-                logits = logits.detach().cpu().numpy()
-                y = y.detach().cpu().numpy()
-                
-                # Track accuracy
-                feat_test.append(feat)
-                logits_test.append(logits)
-                y_test.append(y)
-                self.metric.calculate_batch(feat, logits, y)
-                
-                for cb in self.callbacks:
-                    cb.on_batch_tested(i, feat)
-                
-        for cb in self.callbacks:
-            cb.on_after_test()
-        return np.concatenate(feat_test), np.concatenate(y_test)
-
-    def on_before_train(self, checkpoint):
-        if checkpoint is not None:
-            self.best_metric = checkpoint['accuracy']
-    
-    def on_before_epoch(self, epoch):
-        if self.fit_metric:
-            self.feat_train, self.y_train = [], []
-        
-    def on_after_gradients(self, epoch, ibatch, feat, logits, y, loss):
-        if self.fit_metric:
-            self.feat_train.append(feat.float().detach().cpu().numpy())
-            self.y_train.append(y.detach().cpu().numpy())
-    
-    def on_after_epoch(self, epoch, model, loss_fn, optim, mean_loss):
-        if self.fit_metric:
-            feat_train = np.concatenate(self.feat_train)
-            y_train = np.concatenate(self.y_train)
-            self.metric.fit(feat_train, y_train)
-        feat_test, y_test = self._eval(model)
-        metric_value = self.metric.get()
-        print(f"--------------- Epoch {epoch:02d} Results ---------------")
-        print(f"Test Metric: {metric_value:.6f}")
-        if self.best_epoch != -1:
-            print(f"Best until now: {self.best_metric:.6f}, at epoch {self.best_epoch}")
-        print("------------------------------------------------")
-        if metric_value > self.best_metric:
-            self.best_metric = metric_value
-            self.best_epoch = epoch
-            print('New Best Test Metric!')
-            for cb in self.callbacks:
-                cb.on_best_accuracy(epoch, model, loss_fn, optim, metric_value, feat_test, y_test)
-
-
 class DeviceMapperTransform:
 
     def __init__(self, device):
