+++ metrics.py
@@ -1,3 +1,4 @@
+import torch
 import numpy as np
 from sklearn.neighbors import KNeighborsClassifier
 from scipy.stats import spearmanr
@@ -7,7 +8,7 @@ from pyannote.metrics.binary_classification import det_curve
 from pyannote.core.utils.distance import cdist
 from pyannote.core import Timeline
 from distances import Distance
-from models import SimNet
+from losses.base import TrainingListener
 
 
 class Metric:
@@ -89,7 +90,9 @@ class DistanceSpearmanMetric(Metric):
         pass
 
     def calculate_batch(self, embeddings, logits, y):
-        self.similarity.extend([-self.distance.dist(emb1, emb2) for emb1, emb2 in embeddings])
+        embeddings1, embeddings2 = embeddings
+        self.similarity.extend([-self.distance.dist(embeddings1[i,:].unsqueeze(0), embeddings2[i,:].unsqueeze(0))
+                                for i in range(embeddings1.size(0))])
         self.targets.extend(list(y))
 
     def get(self):
@@ -107,7 +110,7 @@ class SpeakerValidationConfig:
         self.duration = duration
 
 
-class EERMetric(Metric):
+class SpeakerVerificationEvaluator(TrainingListener):
 
     @staticmethod
     def get_hash(trial_file):
@@ -119,27 +122,21 @@ class EERMetric(Metric):
             segments = (try_with,)
         return hash((uri, segments))
 
-    def __init__(self, model: SimNet, device: str, batch_size: int,
-                 distance: Distance, config: SpeakerValidationConfig):
-        self.model = model.to_prediction_model()
+    def __init__(self, device: str, batch_size: int, distance: Distance, config: SpeakerValidationConfig, callbacks=[]):
+        super(SpeakerVerificationEvaluator, self).__init__()
         self.device = device
         self.batch_size = batch_size
         self.distance = distance
         self.config = config
+        self.callbacks = callbacks
+        self.best_metric, self.best_epoch = 0, -1
 
-    def fit(self, embeddings, y):
-        pass
-
-    def calculate_batch(self, embeddings, logits, y):
-        pass
-
-    def get(self):
-        # TODO refactor, maybe transform into a subclass of Evaluator
-        # initialize embedding extraction
-        sequence_embedding = SequenceEmbedding(model=self.model,
+    def _eval(self, model):
+        model.eval()
+        sequence_embedding = SequenceEmbedding(model=model,
                                                feature_extraction=self.config.feature_extraction,
                                                duration=self.config.duration,
-                                               step=.5*self.config.duration,
+                                               step=.5 * self.config.duration,
                                                batch_size=self.batch_size,
                                                device=self.device)
         protocol = get_protocol(self.config.protocol_name, progress=False, preprocessors=self.config.preprocessors)
@@ -176,3 +173,158 @@ class EERMetric(Metric):
         _, _, _, eer = det_curve(np.array(y_true), np.array(y_pred), distances=True)
         # Returning 1-eer because the evaluator keeps track of the highest metric value
         return 1 - eer
+
+    def on_before_train(self, checkpoint):
+        if checkpoint is not None:
+            self.best_metric = checkpoint['accuracy']
+
+    def on_after_epoch(self, epoch, model, loss_fn, optim, mean_loss):
+        metric_value = self._eval(model)
+        print(f"--------------- Epoch {epoch:02d} Results ---------------")
+        print(f"Dev EER: {1 - metric_value:.6f}")
+        if self.best_epoch != -1:
+            print(f"Best until now: {1 - self.best_metric:.6f}, at epoch {self.best_epoch}")
+        print("------------------------------------------------")
+        if metric_value > self.best_metric:
+            self.best_metric = metric_value
+            self.best_epoch = epoch
+            print('New Best Dev EER!')
+            for cb in self.callbacks:
+                cb.on_best_accuracy(epoch, model, loss_fn, optim, metric_value, None, None)
+
+
+class ClassAccuracyEvaluator(TrainingListener):
+
+    def __init__(self, device, loader, metric, batch_transforms=[], callbacks=[]):
+        super(ClassAccuracyEvaluator, self).__init__()
+        self.device = device
+        self.loader = loader
+        self.metric = metric
+        self.batch_transforms = batch_transforms
+        self.callbacks = callbacks
+        self.feat_train, self.y_train = None, None
+        self.best_metric, self.best_epoch = 0, -1
+
+    def _eval(self, model):
+        model.eval()
+        feat_test, logits_test, y_test = [], [], []
+        for cb in self.callbacks:
+            cb.on_before_test()
+        with torch.no_grad():
+            for i in range(self.loader.nbatches()):
+                x, y = next(self.loader)
+
+                # Apply custom transformations to the batch before feeding the model
+                for transform in self.batch_transforms:
+                    x, y = transform(x, y)
+
+                # Feed Forward
+                feat, logits = model(x, y)
+                feat = feat.detach().cpu().numpy()
+                logits = logits.detach().cpu().numpy()
+                y = y.detach().cpu().numpy()
+
+                # Track accuracy
+                feat_test.append(feat)
+                logits_test.append(logits)
+                y_test.append(y)
+                self.metric.calculate_batch(feat, logits, y)
+
+                for cb in self.callbacks:
+                    cb.on_batch_tested(i, feat)
+
+        for cb in self.callbacks:
+            cb.on_after_test()
+        return np.concatenate(feat_test), np.concatenate(y_test)
+
+    def on_before_train(self, checkpoint):
+        if checkpoint is not None:
+            self.best_metric = checkpoint['accuracy']
+
+    def on_before_epoch(self, epoch):
+        self.feat_train, self.y_train = [], []
+
+    def on_after_gradients(self, epoch, ibatch, feat, logits, y, loss):
+        self.y_train.append(y.detach().cpu().numpy())
+
+    def on_after_epoch(self, epoch, model, loss_fn, optim, mean_loss):
+        feat_train = np.concatenate(self.feat_train)
+        y_train = np.concatenate(self.y_train)
+        self.metric.fit(feat_train, y_train)
+        feat_test, y_test = self._eval(model)
+        metric_value = self.metric.get()
+        print(f"--------------- Epoch {epoch:02d} Results ---------------")
+        print(f"Dev Accuracy: {metric_value:.6f}")
+        if self.best_epoch != -1:
+            print(f"Best until now: {self.best_metric:.6f}, at epoch {self.best_epoch}")
+        print("------------------------------------------------")
+        if metric_value > self.best_metric:
+            self.best_metric = metric_value
+            self.best_epoch = epoch
+            print('New Best Dev Accuracy!')
+            for cb in self.callbacks:
+                cb.on_best_accuracy(epoch, model, loss_fn, optim, metric_value, feat_test, y_test)
+
+
+class STSEvaluator(TrainingListener):
+
+    def __init__(self, device, loader, metric, batch_transforms=[], callbacks=[]):
+        super(STSEvaluator, self).__init__()
+        self.device = device
+        self.loader = loader
+        self.metric = metric
+        self.batch_transforms = batch_transforms
+        self.callbacks = callbacks
+        self.best_metric, self.best_epoch = 0, -1
+
+    def _eval(self, model):
+        model.eval()
+        feat_test, y_test = [], []
+        for cb in self.callbacks:
+            cb.on_before_test()
+        with torch.no_grad():
+            for i in range(self.loader.nbatches()):
+                x, y = next(self.loader)
+
+                # Apply custom transformations to the batch before feeding the model
+                for transform in self.batch_transforms:
+                    x, y = transform(x, y)
+
+                # Feed Forward
+                feat = model(x)
+
+                # In evaluation mode, we always receive 2 phrases and no logits
+                feat1 = feat[0].detach().cpu().numpy()
+                feat2 = feat[1].detach().cpu().numpy()
+                y = y.detach().cpu().numpy()
+
+                # Track accuracy
+                feat_test.append((feat1, feat2))
+                y_test.append(y)
+                self.metric.calculate_batch(feat, None, y)
+
+                for cb in self.callbacks:
+                    cb.on_batch_tested(i, feat)
+
+        for cb in self.callbacks:
+            cb.on_after_test()
+        return feat_test, y_test
+
+    def on_before_train(self, checkpoint):
+        if checkpoint is not None:
+            self.best_metric = checkpoint['accuracy']
+
+    def on_after_epoch(self, epoch, model, loss_fn, optim, mean_loss):
+        feat_test, y_test = self._eval(model.to_prediction_model())
+        metric_value = self.metric.get()
+        print(f"--------------- Epoch {epoch:02d} Results ---------------")
+        print(f"Dev Spearman: {metric_value:.6f}")
+        if self.best_epoch != -1:
+            print(f"Best until now: {self.best_metric:.6f}, at epoch {self.best_epoch}")
+        print("------------------------------------------------")
+        if metric_value > self.best_metric:
+            self.best_metric = metric_value
+            self.best_epoch = epoch
+            print('New Best Dev Spearman!')
+            for cb in self.callbacks:
+                cb.on_best_accuracy(epoch, model, loss_fn, optim, metric_value, feat_test, y_test)
